+++ src/client/managers/widgets.js
@@ -19,6 +19,9 @@ class WidgetManager extends EventEmitter {
 
         this.preArgsSeparator = '||||'
 
+        this.valueChangeLockDepth = 0
+        this.valueChangeQueue = {}
+
         this.on('value-changed', this.onChange.bind(this))
 
         ipc.on('connect', ()=>{
@@ -29,7 +32,28 @@ class WidgetManager extends EventEmitter {
 
     }
 
+    valueChangeLock() {
+        this.valueChangeLockDepth++
+    }
+
+    valueChangeUnlock() {
+        this.valueChangeLockDepth--
+        if (this.valueChangeLockDepth == 0) {
+            for (var h in this.valueChangeQueue) {
+                if (this.valueChangeQueue[h]) this.applyValueChange(this.valueChangeQueue[h])
+            }
+            this.valueChangeQueue = {}
+        }
+    }
+
     onChange(e) {
+
+        if (this.valueChangeLockDepth === 0) this.applyValueChange(e)
+        else this.valueChangeQueue[e.widget.hash] = e
+
+    }
+
+    applyValueChange(e) {
 
         var {id, widget, linkId, options} = e
 
@@ -78,6 +102,7 @@ class WidgetManager extends EventEmitter {
             }
         }
 
+        this.trigger('value-changed-lazy', e)
     }
 
     createAddressRef(widget, preArgs, address) {