+++ src/client/parser.js
@@ -43,6 +43,8 @@ var Parser = class Parser {
 
         var props = data
 
+        widgetManager.valueChangeLock()
+
         // Set default widget type
         props.type =  tab ? 'tab' : props.type || 'fader'
 
@@ -106,6 +108,8 @@ var Parser = class Parser {
 
         parentNode.appendChild(widget.container)
         widget.mounted = true
+
+        widgetManager.valueChangeUnlock()
 
         // Editor needs to get the container object
         return widget