+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 1.25.8
 
 - bug fixes
-    - scripting: `getProp()` now correctly returns a copy of the property if it is an object 
+    - scripting: `getProp()` now correctly returns a copy of the property if it is an object
+    - prevent infinite loop when loading widgets with complex inter-dependencies
+
 
 ## 1.25.7